+++ src/main/java/komodocrypto/configuration/exchange_utils/BittrexUtil.java
@@ -1,0 +1,4 @@
+package komodocrypto.configuration.exchange_utils;
+
+public class BittrexUtil {
+}
