+++ src/main/java/komodocrypto/services/exchanges/bitstamp/BitstampAccount.java
@@ -1,18 +1,16 @@
-package komodocrypto.model.exchanges.bitstamp;
+package komodocrypto.services.exchanges.bitstamp;
 
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.util.List;
+
+import komodocrypto.configuration.exchange_utils.BitstampUtil;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.bitstamp.dto.account.BitstampBalance;
-import org.knowm.xchange.bitstamp.dto.account.BitstampDepositAddress;
-import org.knowm.xchange.bitstamp.dto.account.BitstampWithdrawal;
-import org.knowm.xchange.bitstamp.dto.account.DepositTransaction;
-import org.knowm.xchange.bitstamp.dto.account.WithdrawalRequest;
-import org.knowm.xchange.bitstamp.service.BitstampAccountServiceRaw;
 import org.knowm.xchange.currency.Currency;
 import org.knowm.xchange.dto.account.AccountInfo;
+import org.knowm.xchange.dto.account.FundingRecord;
 import org.knowm.xchange.service.account.AccountService;
+import org.knowm.xchange.service.trade.params.TradeHistoryParams;
 
 /**
  * Example showing the following:
@@ -44,6 +42,16 @@ public class BitstampAccount {
 
         String depositAddress = accountService.requestDepositAddress(Currency.BTC);
         System.out.println("Deposit address: " + depositAddress);
+
+        TradeHistoryParams tradeHistoryParams = accountService.createFundingHistoryParams();
+        List<FundingRecord> fundingRecords = accountService.getFundingHistory(tradeHistoryParams);
+        // Only works if you have transaction history. I do not.
+        for (FundingRecord record : fundingRecords) {
+            System.out.println(record.getStatus());
+            System.out.println(record.getBlockchainTransactionHash());
+            System.out.println(record.getAddress());
+            System.out.println(record.getInternalId());
+        }
 
         // WARNING: WILL WITHDRAW COINS IF THERE IS AN ADDRESS
         // Address is set to "XXX" so will throw an error, but don't actually use this to withdraw money.
