+++ src/main/java/komodocrypto/services/exchanges/bitstamp/BitstampTradeImpl.java
@@ -1,4 +1,4 @@
-package komodocrypto.model.exchanges.bitstamp;
+package komodocrypto.services.exchanges.bitstamp;
 
 /**
  *
