+++ src/main/java/komodocrypto/services/exchanges/bitstamp/api-specifications.txt
