+++ src/main/java/komodocrypto/services/exchanges/bittrex/api-specifications.txt
